requirements.txt
python-dotenv       1.0.1
python-telegram-bot 13.0
requests            2.31.0
telegram            0.0.1
urllib3             1.26.15
